MoneyMorpher v0.1.1

+++ src/main/java/com/ada/moneymorpher/Transaction/TransactionService.java
@@ -3,7 +3,9 @@ package com.ada.moneymorpher.Transaction;
 
 import com.ada.moneymorpher.currency.Currency;
 import com.ada.moneymorpher.currency.CurrencyRepository;
+import com.ada.moneymorpher.exceptions.ForbiddenException;
 import com.ada.moneymorpher.exceptions.NotFoundException;
+import com.ada.moneymorpher.profile.ProfileService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
@@ -11,7 +13,6 @@ import org.springframework.stereotype.Service;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -20,6 +21,7 @@ public class TransactionService {
 
     private final TransactionRepository repository;
     private final CurrencyRepository currencyRepository;
+    private final ProfileService profileService;
     private final ModelMapper modelMapper;
 
     private TransactionDto convertDto(Transaction transaction){
@@ -44,7 +46,7 @@ public class TransactionService {
     }
 
 
-    public TransactionDto create(TransactionRequest request){
+    public TransactionDto create(TransactionRequest request, String username){
         Transaction transaction = new Transaction();
         transaction.setUuid(UUID.randomUUID());
         transaction.setDescription(request.getDescription());
@@ -61,11 +63,14 @@ public class TransactionService {
         transaction.setEURValue(EURValue);
         transaction.setUSDValue(USDValue);
 
+        final var profile = this.profileService.getByUsernameEntity(username);
+        transaction.setProfile(profile);
+
         final var saved = this.repository.save(transaction);
         return this.convertDto(saved);
     }
 
-    public List<TransactionList> listTransactions(CurrencyTypeEnum currency) {
+    public List<TransactionList> listAllTransactions(CurrencyTypeEnum currency) {
         List<Transaction> transactionList = this.repository.findAllByOrderByCreatedAtDesc();
 
         return transactionList.stream()
@@ -73,15 +78,23 @@ public class TransactionService {
                 .toList();
     }
 
-    public BigDecimal listBalance(CurrencyTypeEnum currency){
+    public List<TransactionList> listTransactions(CurrencyTypeEnum currency, String username) {
+        List<Transaction> transactionList = this.repository.findByProfileUsernameOrdered(username);
+
+        return transactionList.stream()
+                .map((transaction -> convertList(transaction,currency)))
+                .toList();
+    }
+
+    public BigDecimal listBalance(CurrencyTypeEnum currency, String username){
         BigDecimal balance;
 
         if(currency.equals(CurrencyTypeEnum.USD)){
-            balance = this.repository.getUSDBalance();
+            balance = this.repository.getUSDBalance(username);
         } else if (currency.equals(CurrencyTypeEnum.EUR)){
-            balance = this.repository.getEURBalance();
+            balance = this.repository.getEURBalance(username);
         } else {
-            balance = this.repository.getBRLBalance();
+            balance = this.repository.getBRLBalance(username);
         }
 
         if(balance == null){
@@ -90,21 +103,43 @@ public class TransactionService {
         return balance;
     }
 
-    public Optional<TransactionDto> listDetails (UUID uuid){
-        return this.repository.findByUuid(uuid).map(this::convertDto);
-    }
-
-    public TransactionDto update(UUID uuid, String description){
+    public TransactionDto listDetails (UUID uuid, String username){
         Transaction transaction = this.repository.findByUuid(uuid)
                 .orElseThrow(()-> new NotFoundException("Transaction not found"));
+
+        String transactionUsername = transaction.getProfile().getUsername();
+
+        if(!transactionUsername.equalsIgnoreCase(username)){
+            throw new ForbiddenException("Transaction not allowed to user");
+        }
+
+        return this.convertDto(transaction);
+    }
+
+    public TransactionDto update(UUID uuid, String description, String username){
+        Transaction transaction = this.repository.findByUuid(uuid)
+                .orElseThrow(()-> new NotFoundException("Transaction not found"));
+
+        String transactionUsername = transaction.getProfile().getUsername();
+
+        if(!transactionUsername.equalsIgnoreCase(username)){
+            throw new ForbiddenException("Transaction not allowed to user");
+        }
+
         transaction.setDescription(description);
         Transaction updated = this.repository.save(transaction);
         return this.convertDto(updated);
     }
 
-    public void delete(UUID uuid){
+    public void delete(UUID uuid, String username){
         Transaction transaction = this.repository.findByUuid(uuid)
                 .orElseThrow(()-> new NotFoundException("Transaction not found"));
+
+        String transactionUsername = transaction.getProfile().getUsername();
+
+        if(!transactionUsername.equalsIgnoreCase(username)){
+            throw new ForbiddenException("Transaction not allowed to user");
+        }
         this.repository.delete(transaction);
     }
 }
